requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
pyyaml==6.0.1
plotly==5.20.0
openpyxl==3.1.2
